pyproject.toml
[tool.poetry]
name = "hawkeye-backend"
version = "0.1.0"
description = ""
authors = ["poipiii <nanosonico1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "3.0.2"
pillow = "10.2.0"
onnxruntime = "^1.17.3"
opencv-python = "^4.9.0.80"
numpy = "^1.26.4"
flask-cors = "^4.0.0"
lmdb = "^1.4.1"
einops = "^0.6.1"
pytorch-optimizer = "^2.11.1"
scipy = "^1.11.1"
PyYAML = "^6.0.1"
tb-nightly = "^2.14.0a20230802"
tqdm = "^4.65.0"
opencv-python-headless = "^4.8.1.78"
psutil = "^5.9.8"
pywavelets = "^1.5.0"
streamlit = "^1.33.0"

[tool.ruff]
select = ["ALL"]
ignore = [ "E501", "T201", "D", "CPY001",
	   "TD002", "TD003", "COM812", "ISC001",
	   "N8", "ANN", "S101" ]
target-version = "py311"
preview = true

[tool.ruff.format]
skip-magic-trailing-comma = true
line-ending = "lf"

[tool.ruff.lint.isort]
split-on-trailing-comma = false


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
